provide option -B to bin/build.sh which determines whether to bundle gems to sync Gemfile.lock, by default the script bundles

+++ bin/build.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
-docker run  -v $(pwd):/usr/src/app/ -it --rm ruby sh -c 'cd /usr/src/app/
-bundle config mirror.https://rubygems.org https://gems.ruby-china.com
-bundle install'
+BUNDLE=true
+while getopts ":B" opt; do
+  case ${opt} in
+    B)
+      BUNDLE=false
+      ;;
+    \?) echo "Usage: ./build.sh [-B]"
+      exit
+      ;;
+  esac
+done
+
+if [[ $BUNDLE == 'true' ]]; then
+  docker run  -v $(pwd):/usr/src/app/ -it --rm ruby sh -c 'cd /usr/src/app/
+  bundle config mirror.https://rubygems.org https://gems.ruby-china.com
+  bundle install'
+fi
 docker build -t rails-pangu -f Dockerfile .
